:recycle: refactor: Movendo os tipos do usuário, do post e do email_address para arquivos separados.
