+++ clerk-blog/src/@types/email-address.d.ts
@@ -1,0 +1,6 @@
+export interface EmailAddress {
+  id: string;
+  email_address: string;
+  object: string;
+  userId: string;
+}
